+++ lists/mirror-tv/Post.js
@@ -302,6 +302,7 @@ module.exports = {
 
             await generateSource(existingItem, resolvedData)
 
+
             return resolvedData
         },
         validateInput: async ({